# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI SRE Copilot Application Gateway",
    "description": "Application gateway connecting SRE agents, MCP servers, and persistent case files.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "List all incidents",
        "description": "Returns a list of all active (non-archived) SRE incident case files.",
        "operationId": "list_incidents_api_v1_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "type": "array",
                  "title": "Response List Incidents Api V1 Incidents Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Create a new incident and trigger investigation",
        "description": "Intakes a raw SRE alert, initializes a new incident case file in status NEW, and schedules the multi-agent investigation workflow in the background.",
        "operationId": "create_incident_api_v1_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload / inputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/incidents": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "List all incidents",
        "description": "Returns a list of all active (non-archived) SRE incident case files.",
        "operationId": "list_incidents_api_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "type": "array",
                  "title": "Response List Incidents Api Incidents Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Create a new incident and trigger investigation",
        "description": "Intakes a raw SRE alert, initializes a new incident case file in status NEW, and schedules the multi-agent investigation workflow in the background.",
        "operationId": "create_incident_api_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload / inputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetailResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Structured machine-readable error token (e.g. INCIDENT_NOT_FOUND)."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Detailed human-readable explanation of what went wrong."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual metadata or sub-errors."
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Trace identifier of the failed request."
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO-8601 UTC timestamp when the error was caught."
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message",
          "request_id",
          "timestamp"
        ],
        "title": "ErrorDetailResponse",
        "description": "Global standardized SRE API error response schema.",
        "example": {
          "details": {
            "incident_id": "INC-GHOST"
          },
          "error_code": "INCIDENT_NOT_FOUND",
          "message": "Incident INC-GHOST was not found in persistence.",
          "request_id": "4f17c982-8de3-4697-8267-eb3ec27fadcd",
          "timestamp": "2026-06-27T10:12:15.123456+00:00"
        }
      },
...
      "IncidentCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Optional human-readable title of the incident.",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Optional detailed description of the incident.",
            "default": ""
          },
          "environment": {
            "type": "string",
            "title": "Environment",
            "description": "Target deployment environment (e.g. production, staging).",
            "default": "production"
          },
          "raw_alert": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Alert",
            "description": "The raw alert payload from Prometheus, Alertmanager, etc. Must contain 'name' and 'severity' keys."
          }
        },
        "type": "object",
        "title": "IncidentCreateRequest",
        "description": "Payload to create and trigger a new incident investigation workflow.",
        "example": {
          "description": "Checkout database connection pool usage is at 95%",
          "environment": "production",
          "raw_alert": {
            "alert_id": "AL-12345",
            "annotations": {
              "description": "Database pool connections exhausted.",
              "runbook_url": "https://wiki.company.internal/runbooks/db-pool",
              "summary": "Checkout DB Connection Pool Full"
            },
            "name": "DatabaseDegradation",
            "service": "checkout-db",
            "severity": "P1",
            "started_at": "2026-06-27T10:00:00Z",
            "status": "firing"
          },
          "title": "Database degradation alert"
        }
      },
      "IncidentResponse": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id",
            "description": "Unique incident ID (INC-<8 uppercase hex characters>)."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Human-readable title of the incident."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Detailed description of the incident."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current status in the lifecycle (e.g. INVESTIGATING)."
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "Severity level of the incident (P0-P4)."
          },
          "environment": {
            "type": "string",
            "title": "Environment",
            "description": "Deployment environment."
          },
          "assigned_team": {
            "type": "string",
            "title": "Assigned Team",
            "description": "On-call team assigned to the incident."
          },
          "recovery_status": {
            "type": "string",
            "title": "Recovery Status",
            "description": "Status of recovery actions (e.g. PENDING_APPROVAL)."
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status",
            "description": "Status of recovery verification tests."
          },
          "report_status": {
            "type": "string",
            "title": "Report Status",
            "description": "Status of final report generation."
          },
          "escalation_status": {
            "type": "string",
            "title": "Escalation Status",
            "description": "Status of team notifications."
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO-8601 UTC timestamp of creation."
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO-8601 UTC timestamp of last update."
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "One-line summary of the current incident status."
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence",
            "description": "Overall pipeline confidence score (0-100).",
            "default": 0
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntryResponse"
            },
            "type": "array",
            "title": "Timeline",
            "description": "Audit log timeline entries."
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "title",
          "description",
          "status",
          "severity",
          "environment",
          "assigned_team",
          "recovery_status",
          "verification_status",
          "report_status",
          "escalation_status",
          "created_at",
          "updated_at",
          "summary"
        ],
        "title": "IncidentResponse",
        "description": "Client-facing detail representation of an incident.",
        "example": {
          "assigned_team": "Database-SRE",
          "confidence": 85,
          "created_at": "2026-06-27T10:00:00Z",
          "description": "Triggered on service: checkout-db",
          "environment": "production",
          "escalation_status": "SENT",
          "incident_id": "INC-A4F3B2C1",
          "recovery_status": "NOT_STARTED",
          "report_status": "NOT_STARTED",
          "severity": "P1",
          "status": "INVESTIGATING",
          "summary": "Triage complete. High latency confirmed on checkout service database.",
          "timeline": [],
          "title": "Alert: DatabaseDegradation",
          "updated_at": "2026-06-27T10:01:30Z",
          "verification_status": "NOT_STARTED"
        }
      },
...
      "TimelineEntryResponse": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO-8601 UTC timestamp when the event occurred."
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name",
            "description": "Name of the agent or system that created this entry."
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "Structured type of the event (e.g. ROOT_CAUSE_FOUND)."
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Short verb describing what was executed."
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Human-readable description of the event."
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence",
            "description": "Agent confidence score at this point (0-100)."
          },
          "tools_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools Used",
            "description": "List of MCP tools invoked during this step."
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "description": "Execution duration of this agent step in milliseconds."
          },
          "entry_status": {
            "type": "string",
            "title": "Entry Status",
            "description": "Outcome status of the step (SUCCESS, FAILURE, etc.)."
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "agent_name",
          "event_type",
          "action",
          "summary",
          "confidence",
          "duration_ms",
          "entry_status"
        ],
        "title": "TimelineEntryResponse",
        "description": "Client-facing representation of a timeline event.",
        "example": {
          "action": "rca_completed",
          "agent_name": "RootCauseAgent",
          "confidence": 90,
          "duration_ms": 1250,
          "entry_status": "SUCCESS",
          "event_type": "ROOT_CAUSE_FOUND",
          "summary": "Database connection pool exhaustion detected due to slow transactional queries.",
          "timestamp": "2026-06-27T10:02:15Z",
          "tools_used": [
            "query_logs",
            "get_metrics"
          ]
        }
      },
...
    }
  }
}